Rename Python package from `grafana_api` to `grafana-client`
