+++ .gitignore
@@ -1,5 +1,5 @@
 #
-grafana_api/version.py
+grafana_client/version.py
 
 # Created by .ignore support plugin (hsz.mobi)
 ### Python template
