+++ README.md
@@ -1,6 +1,6 @@
-# grafana_api [![Github Actions Test](https://github.com/m0nhawk/grafana_api/workflows/Test/badge.svg)](https://github.com/m0nhawk/grafana_api/actions?query=workflow%3ATest) [![GitHub license](https://img.shields.io/github/license/m0nhawk/grafana_api.svg?style=flat-square)](https://github.com/m0nhawk/grafana_api/blob/master/LICENSE)  [![Codecov](https://img.shields.io/codecov/c/gh/m0nhawk/grafana_api.svg?style=flat-square)](https://codecov.io/gh/m0nhawk/grafana_api/)
+# grafana-client [![Github Actions Test](https://github.com/m0nhawk/grafana_api/workflows/Test/badge.svg)](https://github.com/m0nhawk/grafana_api/actions?query=workflow%3ATest) [![GitHub license](https://img.shields.io/github/license/m0nhawk/grafana_api.svg?style=flat-square)](https://github.com/m0nhawk/grafana_api/blob/master/LICENSE)  [![Codecov](https://img.shields.io/codecov/c/gh/m0nhawk/grafana_api.svg?style=flat-square)](https://codecov.io/gh/m0nhawk/grafana_api/)
 
-[![PyPI](https://img.shields.io/pypi/v/grafana_api.svg?style=flat-square)](https://pypi.org/project/grafana-api/) [![Conda](https://img.shields.io/conda/v/m0nhawk/grafana_api.svg?style=flat-square)](https://anaconda.org/m0nhawk/grafana_api)
+[![PyPI](https://img.shields.io/pypi/v/grafana-client.svg?style=flat-square)](https://pypi.org/project/grafana-api/) [![Conda](https://img.shields.io/conda/v/m0nhawk/grafana_api.svg?style=flat-square)](https://anaconda.org/m0nhawk/grafana_api)
 
 ## What is this library for?
 
@@ -15,39 +15,39 @@ You need Python 3 and only the `requests` library installed.
 Install the pip package:
 
 ```
-pip install -U grafana_api
+pip install -U grafana-client
 ```
 
 And then connect to your Grafana API endpoint:
 
 ```python
-from grafana_api.grafana_face import GrafanaFace
+from grafana_client.grafana_face import GrafanaFace
 
-grafana_api = GrafanaFace(auth='abcde....', host='api.my-grafana-host.com')
+grafana = GrafanaFace(auth='abcde....', host='api.my-grafana-host.com')
 
 # Create user
-user = grafana_api.admin.create_user({"name": "User", "email": "user@domain.com", "login": "user", "password": "userpassword", "OrgId": 1})
+user = grafana.admin.create_user({"name": "User", "email": "user@domain.com", "login": "user", "password": "userpassword", "OrgId": 1})
 
 # Change user password
-user = grafana_api.admin.change_user_password(2, "newpassword")
+user = grafana.admin.change_user_password(2, "newpassword")
 
 # Search dashboards based on tag
-grafana_api.search.search_dashboards(tag='applications')
+grafana.search.search_dashboards(tag='applications')
 
 # Find a user by email
-user = grafana_api.users.find_user('test@test.com')
+user = grafana.users.find_user('test@test.com')
 
 # Add user to team 2
-grafana_api.teams.add_team_member(2, user["id"])
+grafana.teams.add_team_member(2, user["id"])
 
 # Create or update a dashboard
-grafana_api.dashboard.update_dashboard(dashboard={'dashboard': {...}, 'folderId': 0, 'overwrite': True})
+grafana.dashboard.update_dashboard(dashboard={'dashboard': {...}, 'folderId': 0, 'overwrite': True})
 
 # Delete a dashboard by UID
-grafana_api.dashboard.delete_dashboard(dashboard_uid='abcdefgh')
+grafana.dashboard.delete_dashboard(dashboard_uid='abcdefgh')
 
 # Create organization
-grafana_api.organization.create_organization({"name":"new_organization"})
+grafana.organization.create_organization({"name":"new_organization"})
 ```
 
 
@@ -60,13 +60,13 @@ To use admin API you need to use basic auth [as stated here](https://grafana.com
 ```python
 # Use basic authentication:
 
-grafana_api = GrafanaFace(
+grafana = GrafanaFace(
           auth=("username","password"),
           host='api.my-grafana-host.com'
           )
 
 # Use token
-grafana_api = GrafanaFace(
+grafana = GrafanaFace(
           auth='abcdetoken...',
           host='api.my-grafana-host.com'
           )
@@ -100,7 +100,7 @@ Work on API implementation still in progress.
 
 ## Issue tracker
 
-Please report any bugs and enhancement ideas using the `grafana_api` issue tracker:
+Please report any bugs and enhancement ideas using the `grafana-client` issue tracker:
 
   https://github.com/m0nhawk/grafana_api/issues
 
@@ -108,5 +108,5 @@ Feel free to also ask questions on the tracker.
 
 ## License
 
-`grafana_api` is licensed under the terms of the MIT License (see the file
+`grafana-client` is licensed under the terms of the MIT License (see the file
 [LICENSE](LICENSE)).
