+++ grafana_client/api/admin.py
