+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = grafana_api
+name = grafana-client
 friendly_name = Python Grafana API
 description = Yet another Python library for Grafana API
 long_description = file: README.md
@@ -57,4 +57,4 @@ universal = true
 [tool.setuptools_scm]
 local_scheme = no-local-version
 version_scheme = python-simplified-semver
-write_to = grafana_api/version.py
+write_to = grafana_client/version.py
