+++ test/api/test_annotations.py
@@ -2,10 +2,10 @@ import unittest
 
 import requests_mock
 
-from grafana_api.grafana_api import (GrafanaBadInputError, GrafanaClientError,
+from grafana_client.grafana_api import (GrafanaBadInputError, GrafanaClientError,
                                      GrafanaServerError,
                                      GrafanaUnauthorizedError)
-from grafana_api.grafana_face import GrafanaFace
+from grafana_client.grafana_face import GrafanaFace
 
 
 class AnnotationsTestCase(unittest.TestCase):
