+++ test/api/test_search.py
@@ -2,8 +2,8 @@ import unittest
 
 import requests_mock
 
-from grafana_api.grafana_face import GrafanaFace
-from grafana_api.grafana_api import GrafanaServerError, GrafanaClientError, GrafanaUnauthorizedError, \
+from grafana_client.grafana_face import GrafanaFace
+from grafana_client.grafana_api import GrafanaServerError, GrafanaClientError, GrafanaUnauthorizedError, \
     GrafanaBadInputError
