+++ test/api/test_snapshot.py
@@ -2,7 +2,7 @@ import unittest
 
 import requests_mock
 
-from grafana_api.grafana_face import GrafanaFace
+from grafana_client.grafana_face import GrafanaFace
 
 
 class SnapshotTestCase(unittest.TestCase):
